Base classes wrap-up
Updated and cleaned up the base classes in preparation for creating different kinds of towers and enemies.

The most notable addition is functional targeting modes to towers, allowing them to target the first enemy, last enemy, highest hp enemy, highest defense enemy, or fastest enemy. We don't yet have a UI to change this in game, so for now all towers will be on the default targeting (first). With this comes a 'distance travelled' for enemies, which tracks how far along the track they are.

Other changes:
- Piercing should be fixed now and fully functional. A projectile with pierce set to -1 (or any negative value) should in theory have infinite pierce.
- Now only projectiles fired by towers with aerial targeting can collide with aerial enemies. Before, non-aerial towers could still indirectly hit aerial enemies when aiming at grounded ones.
- Towers now only rotate / aim to do an attack, which combined with targeting implementation makes them look a lot less janky in action.