Shop Menu
Merged BuildManager and LevelManager as BuildManager was somewhat redundant.

Created a basic interface for the shop system where the player can open a menu select a tower and build said tower.
The menu can be open and closed using a toggle button, making use of simple animations to move the menu in and out of the screen.

Tied the building of the towers to the in-game currency which is earned when defeating enemies. (Slightly annoying for testing possibly adding a cheat code system / infinite currency toggle would be helpful)

Implemented a temp TestTower2 as a proof of concept for the shop being able to switch between types of tower.
This tower is "blue-ish" and more powerful than the TestTower.

Modified Plot.cs to ensure that towers can not be built through the menu (or other future UI items).
Modified Plot.cs to ensure that tiles can not be "selected" through the menu (or other future UI items).

Modified BuildManager.cs to allow the player to switch between different types of towers / shop items.

Added a Tower.cs file this manages the interpretation of the tower from the shop perspective and deals with the cost of given towers (This may be something that can be passed through by individual towers, but for now this works).