New ranged towers implementation
The main additions in this version are the implementations of all the ranged towers which Michal had finished creating art for.
Towers implemented:
- Archer: Shoots quickly in a medium range, but has no pierce and limited damage. Has aerial targeting. Uses the base tower class - no complex behaviour.
- Ballista: Shoots slowly at long range, with high damange and pierce. Does not have aerial targetting. Uses the base tower class - no complex behaviour.
- Catapult: Shoots slowly at great range, exceptional damage and pierce. Has aerial targetting. Uses a derived class with its key feature being unique targetting modes - the catapult can only fire in a straight line in the cardinal directions, either up, down, left, or right. It will not target enemies not in the straight line it is set to fire down.

In addition I have edited a few base class scripts for their functionality - as well as the Witch enemy. I've also re-organised our files to help with the structure of the project overall.