Enemy Spawning Updated
Enemies now spawn based on a list of integers that correspond to the index of the enemy to be spawned within the enemyPrefabs array.
There is an error where the game will spawn all but the last enemy in the list, however my attempts to fix this cause infinite and rapid spawning of the enemy type (which is obviously not intended).